Refactor view
