--- src/wireframes/components/animation/ErrorView.tsx
@@ -1,9 +1,0 @@
-import * as React from "react";
-
-export class ErrorView extends React.Component {
-    render() {
-        return (
-            <div id="error-view" style={{ height: 200, whiteSpace: 'pre-wrap', padding: '0 20px' }}></div>
-        );
-    }
-}
