+++ src/wireframes/components/layout/EditorView.scss
