+++ src/wireframes/components/layout/PanelView.scss
