+++ src/wireframes/components/layout/PanelView.tsx
@@ -1,10 +1,69 @@
+import { Button, Tabs } from "antd";
 import * as React from "react";
-import './AnimationView.scss';
-import { changeItemsAppearance, getDiagram, selectItems, useStore } from "@app/wireframes/model";
+import './PanelView.scss';
+import { useEventCallback } from '@app/core';
+import { Diagram, changeItemsAppearance, getDiagram, selectItems, selectPanel, togglePanel, useStore } from "@app/wireframes/model";
 import { useDispatch } from "react-redux";
 import { useState } from "react";
+import { CloseOutlined } from "@ant-design/icons";
+import { texts } from "@app/texts";
 
-export const AnimationView = () => {
+export interface PanelViewProps {
+}
+
+export const PanelView = (props: PanelViewProps) => {
+    const diagram = useStore(getDiagram);
+
+    if (!diagram) {
+        return null;
+    }
+
+    return (
+        <PanelViewInner {...props} diagram={diagram} />
+    );
+};
+
+export const PanelViewInner = ({ diagram }: PanelViewProps & { diagram: Diagram }) => {
+    const dispatch = useDispatch();
+    const selectedPanel = useStore(s => s.ui.selectedPanel);
+    const showPanel = useStore(s => s.ui.showPanel);
+    
+    const doSelectPanel = useEventCallback((key: string) => {
+        dispatch(selectPanel(key));
+        if (!showPanel) {
+            doTogglePanel();
+        }
+    });
+    const doTogglePanel = useEventCallback(() => {
+        dispatch(togglePanel());
+    });
+    
+    const closePanel = <Button onClick={doTogglePanel}
+        icon={showPanel ? <CloseOutlined /> : <></>}
+        type='text'
+        size='small'
+        shape='circle' />
+
+
+    return (
+        <Tabs tabBarExtraContent={closePanel} className="animation-view" type='card' onTabClick={doSelectPanel} activeKey={selectedPanel}>
+            <Tabs.TabPane key='animation' tab={texts.common.animation}>
+                { showPanel ? <AnimationView /> : <></> }
+            </Tabs.TabPane>
+            <Tabs.TabPane key='error' tab={texts.common.error}>
+                { showPanel ? <ErrorView /> : <></> }
+            </Tabs.TabPane>
+        </Tabs>
+    );
+};
+
+const ErrorView = () => {
+    return (
+        <div id="error-view" style={{ height: 200, whiteSpace: 'pre-wrap', padding: '0 20px' }}></div>
+    );
+};
+
+const AnimationView = () => {
     const dispatch = useDispatch();
     const [code, setCode] = useState('');
     const selectedDiagram = useStore(getDiagram);
@@ -66,5 +125,4 @@ export const AnimationView = () => {
             <textarea value={code} onChange={selectShape} onKeyUp={addLineNumber} onMouseLeave={revertShape}/>
         </div>
     );
-}
-
+};
